Design aside container with logo and cover photo
I designed the aside container with the website's logo, title, cover
photo, and footer links. The website's logo container and footer used
a semi-transparent black background to enhance the readability of the
text.

I also utilized positioning styles to make the logo container and
footer overlay the website's cover photo.

For the footer links, I added the links to give credit for the cover
photo's creator.