Design main container content except form
I designed the content of the main container to follow the styles of the
reference image. I tried my best to make the margins and paddings of the
content to be similar to the reference image.

For the submit button, I set the color to one of the tones found in the
cover photo, I also set its hovered state to have a slightly darker
color and its cursor changed to pointer.

For the login link, I set the color similar to the submit button.